+++ lib/wcc/contentful/redirect.rb
@@ -14,11 +14,19 @@ module WCC::Contentful
 
     def location
       if !self.url.nil?
-        return self.url
-      elsif !self.pageReference.nil?
-        return "/#{self.pageReference.url}"
+        self.url
+      elsif valid_page_reference?(self.pageReference)
+        "/#{self.pageReference.url}"
       else
-        return nil
+        nil
+      end
+    end
+
+    def valid_page_reference?(page_ref)
+      unless page_ref.nil? || defined?(page_ref.url).nil?
+        true
+      else
+        false
       end
     end
   end