+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## v0.0.1
 
 * initial release
+
+## v0.0.2
+
+* Will now return nil if the Redirect model's pageReference does not include a url
+* Add test coverage for the Configuration class
+* Add tests for the valid_page_reference? method